head turning again etc

+++ SnakeGame/SnakeGame/SnakeElem.cs
@@ -5,11 +5,11 @@ namespace SnakeGame
     public class SnakeElem
     {
         private GamepageSnake.Directions direction;
-        private double x, y;
+        private int x, y;
         private Rectangle rect;
 
-        public double X { get => x; set => x = value; }
-        public double Y { get => y; set => y = value; }
+        public int X { get => x; set => x = value; }
+        public int Y { get => y; set => y = value; }
         public Rectangle Rect { get => rect; set => rect = value; }
         public GamepageSnake.Directions Direction { get => direction; set => direction = value; }
     }
